[Refactor]: slices 폴더 생성
- redux관련 파일을 관리하는 slices 폴더를 만들었습니다.
